requirements.txt
--index-url https://download.pytorch.org/whl/cu130
--extra-index-url https://pypi.org/simple
torch==2.10.0
torchvision==0.25.0
numpy>=2.1.0,<3.0.0
pandas>=2.2.0,<3.1.0
matplotlib>=3.9.0,<3.11.0
seaborn>=0.13.0,<0.14.0
scipy>=1.14.0,<1.16.0
scikit-learn>=1.5.0,<1.7.0
Pillow>=11.0.0,<13.0.0
tqdm>=4.66.0,<5.0.0
optuna>=4.3.0,<5.0.0
PyYAML>=6.0.0,<7.0.0
